Move PieRow into metabase-types
