+++ frontend/src/metabase-types/api/visualization-settings.ts
@@ -42,3 +42,14 @@ export type SmartScalarComparison =
   | SmartScalarComparisonPreviousPeriod
   | SmartScalarComparisonPeriodsAgo
   | SmartScalarComparisonStaticNumber;
+
+export interface PieRow {
+  key: string;
+  name: string;
+  originalName: string;
+  color: string;
+  defaultColor: boolean;
+  enabled: boolean;
+  hidden: boolean;
+  isOther: boolean;
+}
