+++ frontend/src/metabase/visualizations/shared/settings/pie.ts
@@ -5,7 +5,6 @@ import { NULL_DISPLAY_VALUE } from "metabase/utils/constants";
 import { checkNotNull, checkNumber, isNumber } from "metabase/utils/types";
 import { SLICE_THRESHOLD } from "metabase/visualizations/echarts/pie/constants";
 import { getPieColumns } from "metabase/visualizations/echarts/pie/model";
-import type { PieRow } from "metabase/visualizations/echarts/pie/model/types";
 import type { ShowWarning } from "metabase/visualizations/echarts/types";
 import { getHexColor } from "metabase/visualizations/lib/color";
 import { getNumberOr } from "metabase/visualizations/lib/settings/row-values";
@@ -18,6 +17,7 @@ import type {
 import {
   type DatasetColumn,
   type MaybeTranslatedSeries,
+  type PieRow,
   type RawSeries,
   type RowValue,
   type RowValues,
