+++ frontend/src/metabase/visualizations/shared/settings/pie.unit.spec.ts
@@ -1,6 +1,5 @@
 import { NULL_DISPLAY_VALUE } from "metabase/utils/constants";
-import type { PieRow } from "metabase/visualizations/echarts/pie/model/types";
-import type { RawSeries } from "metabase-types/api";
+import type { PieRow, RawSeries } from "metabase-types/api";
 import {
   createMockColumn,
   createMockDatasetData,
